Fixed comparisons between signed and unsigned integers.

+++ trc/coffin.cxx
@@ -766,7 +766,8 @@ void Aeon::WriteOutputs() {
                                        icpt < isamp->end(); icpt += 3)
         sampfile << icpt[0] << " " << icpt[1] << " " << icpt[2] << endl;
 
-      if (isamp - mControlPointSamples.begin() == mMaximumAPosterioriPath)
+      if ((int) (isamp - mControlPointSamples.begin())
+           == mMaximumAPosterioriPath)
         cptsmap.insert(cptsmap.end(), isamp->begin(), isamp->end());
 
       if (isamp - mControlPointSamples.begin() == mMaximumAPosterioriPath0)
@@ -790,7 +791,8 @@ void Aeon::WriteOutputs() {
         sampfile << round(point[0]) << " " << round(point[1]) << " "
                                            << round(point[2]) << endl;
 
-        if (isamp - mControlPointSamples.begin() == mMaximumAPosterioriPath)
+        if ((int) (isamp - mControlPointSamples.begin())
+             == mMaximumAPosterioriPath)
           for (int k = 0; k < 3; k++)
             cptsmap.push_back((int) round(point[k]));
